Проект переименован в .Api и изменён в соответствии с новой архитектурой. Тестовый проект изменён в соответствии с новой архитектурой.
