+++ BuildRevisionCounter.Api/Code/Data/Enum/AdminRoles.cs
@@ -1,0 +1,7 @@
+﻿namespace BuildRevisionCounter.Api.Code.Data.Enum
+{
+    public class AdminRoles : DAL.Enums.AdminRoles
+    {
+        public const string AdminEditorAnonymous = "admin, editor, anonymous";
+    }
+}
