+++ BuildRevisionCounter.Api/Properties/AssemblyInfo.cs
@@ -1,0 +1,6 @@
+﻿using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyTitle("BuildRevisionCounter")]
+[assembly: AssemblyVersion("2.0.0.0")]
