+++ BuildRevisionCounter.Api/Security/BasicAuthenticationAttribute.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace BuildRevisionCounter.Api.Security
+{
+	public class BasicAuthenticationAttribute : Attribute
+	{
+	}
+}
