+++ BuildRevisionCounter.Api/packages.config
@@ -1,0 +1,21 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.AspNet.WebApi.Client" version="5.2.3" targetFramework="net451" />
+  <package id="Microsoft.AspNet.WebApi.Core" version="5.2.3" targetFramework="net451" />
+  <package id="Microsoft.AspNet.WebApi.Owin" version="5.2.3" targetFramework="net451" />
+  <package id="Microsoft.Owin" version="3.0.1" targetFramework="net451" />
+  <package id="Microsoft.Owin.Host.SystemWeb" version="3.0.1" targetFramework="net451" />
+  <package id="Modernizr" version="2.6.2" targetFramework="net451" />
+  <package id="MongoDB.Bson" version="2.0.0" targetFramework="net451" />
+  <package id="MongoDB.Driver" version="2.0.0" targetFramework="net451" />
+  <package id="MongoDB.Driver.Core" version="2.0.0" targetFramework="net451" />
+  <package id="Newtonsoft.Json" version="6.0.8" targetFramework="net451" />
+  <package id="Ninject" version="3.2.2.0" targetFramework="net451" />
+  <package id="Ninject.Extensions.ContextPreservation" version="3.2.0.0" targetFramework="net451" />
+  <package id="Ninject.Extensions.NamedScope" version="3.2.0.0" targetFramework="net451" />
+  <package id="Ninject.Web.Common" version="3.2.3.0" targetFramework="net451" />
+  <package id="Ninject.Web.Common.OwinHost" version="3.2.3.0" targetFramework="net451" />
+  <package id="Ninject.Web.WebApi" version="3.2.4.0" targetFramework="net451" />
+  <package id="Ninject.Web.WebApi.OwinHost" version="3.2.4.0" targetFramework="net451" />
+  <package id="Owin" version="1.0" targetFramework="net451" />
+</packages>
