+++ BuildRevisionCounter.Tests/packages.config
@@ -1,7 +1,13 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="Microsoft.Owin" version="3.0.1" targetFramework="net451" />
+  <package id="Microsoft.Owin.Host.HttpListener" version="3.0.1" targetFramework="net451" />
+  <package id="Microsoft.Owin.Hosting" version="3.0.1" targetFramework="net451" />
+  <package id="Microsoft.Owin.Testing" version="3.0.1" targetFramework="net451" />
   <package id="MongoDB.Bson" version="2.0.0" targetFramework="net451" />
   <package id="MongoDB.Driver" version="2.0.0" targetFramework="net451" />
   <package id="MongoDB.Driver.Core" version="2.0.0" targetFramework="net451" />
+  <package id="Moq" version="4.2.1502.0911" targetFramework="net451" />
   <package id="NUnit" version="2.6.4" targetFramework="net451" />
+  <package id="Owin" version="1.0" targetFramework="net451" />
 </packages>
