+++ BuildRevisionCounter.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2013
 VisualStudioVersion = 12.0.31101.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BuildRevisionCounter", "BuildRevisionCounter\BuildRevisionCounter.csproj", "{64EEEA30-FCE1-4AA6-A375-EBD721914708}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BuildRevisionCounter.Api", "BuildRevisionCounter.Api\BuildRevisionCounter.Api.csproj", "{64EEEA30-FCE1-4AA6-A375-EBD721914708}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BuildRevisionCounter.Tests", "BuildRevisionCounter.Tests\BuildRevisionCounter.Tests.csproj", "{837D9E7D-D0CC-4924-85B5-916FC5D931BA}"
 EndProject
