--- BuildRevisionCounter/Security/BasicAuthenticationAttribute.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace BuildRevisionCounter.Security
-{
-	public class BasicAuthenticationAttribute : Attribute
-	{
-	}
-}
