--- BuildRevisionCounter/Web.config
@@ -1,26 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-	<connectionStrings>
-		<add name="MongoDBStorage" connectionString="mongodb://localhost/brCounter" />
-	</connectionStrings>
-	<system.web>
-		<compilation debug="true" targetFramework="4.5.1" />
-		<httpRuntime targetFramework="4.5.1" />
-	</system.web>
-	<system.webServer>
-		<handlers>
-			<remove name="ExtensionlessUrlHandler-Integrated-4.0" />
-			<remove name="OPTIONSVerbHandler" />
-			<remove name="TRACEVerbHandler" />
-			<add name="ExtensionlessUrlHandler-Integrated-4.0" path="*." verb="*" type="System.Web.Handlers.TransferRequestHandler" preCondition="integratedMode,runtimeVersionv4.0" />
-		</handlers>
-	</system.webServer>
-	<runtime>
-		<assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
-			<dependentAssembly>
-				<assemblyIdentity name="Microsoft.Owin" publicKeyToken="31bf3856ad364e35" culture="neutral" />
-				<bindingRedirect oldVersion="0.0.0.0-3.0.0.0" newVersion="3.0.0.0" />
-			</dependentAssembly>
-		</assemblyBinding>
-	</runtime>
-</configuration>
