--- BuildRevisionCounter/packages.config
@@ -1,20 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.AspNet.WebApi.Client" version="5.0.0" targetFramework="net451" />
-  <package id="Microsoft.AspNet.WebApi.Core" version="5.0.0" targetFramework="net451" />
-  <package id="Microsoft.AspNet.WebApi.Owin" version="5.0.0" targetFramework="net451" />
-  <package id="Microsoft.Owin" version="3.0.0" targetFramework="net451" />
-  <package id="Microsoft.Owin.Host.SystemWeb" version="3.0.0" targetFramework="net451" />
-  <package id="MongoDB.Bson" version="2.0.0" targetFramework="net451" />
-  <package id="MongoDB.Driver" version="2.0.0" targetFramework="net451" />
-  <package id="MongoDB.Driver.Core" version="2.0.0" targetFramework="net451" />
-  <package id="Newtonsoft.Json" version="4.5.11" targetFramework="net451" />
-  <package id="Ninject" version="3.2.2.0" targetFramework="net451" />
-  <package id="Ninject.Extensions.ContextPreservation" version="3.2.0.0" targetFramework="net451" />
-  <package id="Ninject.Extensions.NamedScope" version="3.2.0.0" targetFramework="net451" />
-  <package id="Ninject.Web.Common" version="3.2.3.0" targetFramework="net451" />
-  <package id="Ninject.Web.Common.OwinHost" version="3.2.3.0" targetFramework="net451" />
-  <package id="Ninject.Web.WebApi" version="3.2.4.0" targetFramework="net451" />
-  <package id="Ninject.Web.WebApi.OwinHost" version="3.2.4.0" targetFramework="net451" />
-  <package id="Owin" version="1.0" targetFramework="net451" />
-</packages>
